+++ client/src/components/signup/SignupBasicForm/index.tsx
@@ -2,7 +2,7 @@ import { Controller, SubmitHandler, useForm } from 'react-hook-form';
 import { authAPI } from 'apis/authAPI';
 import { AxiosError } from 'axios';
 import { useSetRecoilState } from 'recoil';
-import { useSignupStepControls } from 'routes/Signup';
+import { useSignupStepControls } from 'routes/AuthPage/Signup';
 import AuthFormInput from 'components/common/AuthFormInput';
 import ErrorMessage from 'components/common/ErrorMessage';
 import { getSignupFormFieldErrorMessage } from 'utils/getErrorMessage';
