+++ client/src/routes/AuthPage/Login/Login.module.scss
@@ -1,4 +1,4 @@
-@use '../../styles/variables.scss' as *;
+@use '../../../styles/variables.scss' as *;
 
 .links {
   display: flex;
