+++ client/src/routes/AuthPage/ResetPassword/ResetPassword.module.scss
@@ -1,0 +1,1 @@
+@use '../../../styles/variables.scss' as *;
