+++ client/src/routes/AuthPage/Signup/Signup.module.scss
@@ -1,4 +1,4 @@
-@use '../../styles/variables.scss' as *;
+@use '../../../styles/variables.scss' as *;
 
 .header {
   display: flex;
