--- client/src/routes/FindEmail/FindEmail.module.scss
@@ -1,1 +1,0 @@
-@use '../../styles/variables.scss' as *;
